Fix multicnn with no labels
Former-commit-id: d07b07a111fb5a385f12b951ece159b04a86f360

+++ clinicadl/clinicadl/tools/deep_learning/cnn_utils.py
@@ -343,20 +343,25 @@ def concat_multi_cnn_results(output_dir, fold, selection, mode, dataset, num_cnn
         cnn_metrics_path = os.path.join(performance_dir, '%s_%s_level_metrics.tsv' % (dataset, mode))
 
         cnn_pred_df = pd.read_csv(cnn_pred_path, sep='\t')
-        cnn_metrics_df = pd.read_csv(cnn_metrics_path, sep='\t')
         prediction_df = pd.concat([prediction_df, cnn_pred_df])
-        metrics_df = pd.concat([metrics_df, cnn_metrics_df])
+        os.remove(cnn_pred_path)
+
+        if os.path.exists(cnn_metrics_path):
+            cnn_metrics_df = pd.read_csv(cnn_metrics_path, sep='\t')
+            metrics_df = pd.concat([metrics_df, cnn_metrics_df])
+            os.remove(cnn_metrics_path)
 
         # Clean unused files
-        os.remove(cnn_pred_path)
-        os.remove(cnn_metrics_path)
         if len(os.listdir(performance_dir)) == 0:
             os.rmdir(performance_dir)
         if len(os.listdir(cnn_dir)) == 0:
             os.rmdir(cnn_dir)
 
     prediction_df.reset_index(drop=True, inplace=True)
-    metrics_df.reset_index(drop=True, inplace=True)
+    if len(metrics_df) == 0:
+        metrics_df = None
+    else:
+        metrics_df.reset_index(drop=True, inplace=True)
     mode_level_to_tsvs(output_dir, prediction_df, metrics_df, fold, selection, mode, dataset)
